requirements.txt
APScheduler==3.6.3
astroid==2.6.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
DateTime==4.3
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.9.3
python-telegram-bot==13.7
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
wrapt==1.12.1
zope.interface==5.4.0
